Add cast to ulong

+++ test/csharp/TestOnnxRuntimeGenAIAPI.cs
@@ -589,7 +589,7 @@ namespace Microsoft.ML.OnnxRuntimeGenAI.Tests
 
                     Assert.Equal(50256, tokenizer.GetBosTokenId());
                     Assert.Equal(50256, tokenizer.GetPadTokenId());
-                    Assert.Equal(1, tokenizer.GetEosTokenIds().NumSequences);
+                    Assert.Equal((ulong)1, tokenizer.GetEosTokenIds().NumSequences);
                     Assert.Equal(50256, tokenizer.GetEosTokenIds()[0][0]);
 
                     var strings = new string[] {
